Extract ProjectLoader interface
as it will make easier to create delegates safely.

Bug: n/a
Test: n/a

AOSP: 7fd339f4b8c6f8f5409e01c390ff3bf4455aae79
