+++ base/src/com/google/idea/blaze/base/qsync/QuerySyncManager.java
@@ -161,7 +161,7 @@ public class QuerySyncManager implements Disposable {
   }
 
   protected ProjectLoader createProjectLoader(ListeningExecutorService executor, Project project) {
-    return new ProjectLoader(executor, project);
+    return new ProjectLoaderImpl(executor, project);
   }
 
   public ModificationTracker getProjectModificationTracker() {
